Fonts separated from lcd7920 code. Correctly working.
